% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{binom}
\alias{binom}
\title{Generate n pseudo random numbers from a binomial distribution.}
\usage{
binom(n = 1, p = 0.5, k = 5, seed = as.integer(Sys.time()))
}
\arguments{
\item{n}{the number of pseudo random binomial variants to generate.
Defaults to 1.}

\item{p}{probability of success. Defaults to 0.5.}

\item{k}{Number of n independent Bernoulli trials. Defaults to 5.}

\item{seed}{Seed for pseudo random number generating algorithm. Defaults to a
integer generated from the system time}
}
\value{
vector of binomial random variants of length n
}
\description{
Generate n pseudo random numbers from a binomial distribution.
}
\examples{
binom(n=1e5, p= 0.2, k=5)
}
