% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy.R
\name{cauchy}
\alias{cauchy}
\title{Generate n pseudo random numbers from a cauchy distribution.}
\usage{
cauchy(n = 1, t = 0, s = 1, seed = as.integer(Sys.time()))
}
\arguments{
\item{n}{the number of pseudo random geometric numbers to generate.
Defaults to 1.}

\item{t}{Location parameter. Defaults to 0.}

\item{s}{scaling parameter. Defaults to 1.}

\item{seed}{Seed for pseudo random number generating algorithm. Defaults to a
integer generated from the system time}
}
\value{
vector of cauchy random variants of length n
}
\description{
Generate n pseudo random numbers from a cauchy distribution.
}
\examples{
cauchy(n=1e5, t = -2, s = 2)
}
