% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiSquared.R
\name{chiSquared}
\alias{chiSquared}
\title{Generate n pseudo random numbers from a chi-squared distribution.}
\usage{
chiSquared(n = 1, k = 1, seed = as.integer(Sys.time()))
}
\arguments{
\item{n}{the number of pseudo random  chi-squared numbers to generate.
Defaults to 1.}

\item{k}{degrees is the degrees of freedom. Defaults to 1.}

\item{seed}{Seed for pseudo random number generating algorithm. Defaults to a
integer generated from the system time}
}
\value{
vector of  chi-squared random variants of length n
}
\description{
Generate n pseudo random numbers from a chi-squared distribution.
}
\examples{
chiSquared(n=1e5, k =6, seed = 155)
}
