% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expo.R
\name{expo}
\alias{expo}
\title{Generate n pseudo random numbers from a exponential distribution.}
\usage{
expo(n = 1, lambda = 1, seed = as.integer(Sys.time()))
}
\arguments{
\item{n}{the number of pseudo random exponential numbers to generate.
Defaults to 1.}

\item{lambda}{rate parameter lambda of the distribution. Defaults to 1.}

\item{seed}{Seed for pseudo random number generating algorithm. Defaults to a
integer generated from the system time}
}
\value{
vector of  exponential random variants of length n
}
\description{
Generate n pseudo random numbers from a exponential distribution.
}
\examples{
expo(n=1500, lambda = 6)
}
