% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaSim.R
\name{gammaSim}
\alias{gammaSim}
\title{Generate n pseudo random numbers from a gamma distribution.}
\usage{
gammaSim(n = 1, shape = 2, scale = 1, seed = as.integer(Sys.time()))
}
\arguments{
\item{n}{the number of pseudo random gamma numbers to generate.
Defaults to 1.}

\item{shape}{Shape parameter.  Defaults to 2.}

\item{scale}{scale parameter. Note this is the inverse of the rate parameter
used in some implementations of the gamma function  Defaults to 1.}

\item{seed}{Seed for pseudo random number generating algorithm. Defaults to a
integer generated from the system time}
}
\value{
vector of gamma random variants of length n
}
\description{
Generate n pseudo random numbers from a gamma distribution.
}
\examples{
gammaSim(n=1e5, shape = 3 , scale = 1)
}
