% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric.R
\name{geometric}
\alias{geometric}
\title{Generate n pseudo random numbers from a geometric distribution.}
\usage{
geometric(n = 1, p = 0.5, seed = as.integer(Sys.time()))
}
\arguments{
\item{n}{the number of pseudo random geometric numbers to generate.
Defaults to 1.}

\item{p}{probability of success. Defaults to 0.5.}

\item{seed}{Seed for pseudo random number generating algorithm. Defaults to a
integer generated from the system time}
}
\value{
vector of geometric random variants of length n
}
\description{
Generate n pseudo random numbers from a geometric distribution.
}
\examples{
geometric(n=1e5, p= .3)
}
