% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.R
\name{lognormal}
\alias{lognormal}
\title{Generate n pseudo random numbers from a lognormal distribution with mean 'mu' and
standard deviation 'sd'}
\usage{
lognormal(n = 1, mu = 0, sd = 0.25, seed = as.integer(Sys.time()))
}
\arguments{
\item{n}{the number of pseudo random lognormal numbers to generate.
Defaults to 1.}

\item{mu}{mean of normal distribution. Defaults to 0.}

\item{sd}{standard deviation of normal distribution. Defaults to 0.25.}

\item{seed}{Seed for pseudo random number generating algorithm. Defaults to a
integer generated from the system time}
}
\value{
vector of PRNs from specified normal distribution of length n
}
\description{
Generate n pseudo random numbers from a lognormal distribution with mean 'mu' and
standard deviation 'sd'
}
\examples{
lognormal(n=1e6, mu=0, sd = 3, seed = 123)
}
