% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{normal}
\alias{normal}
\title{Generate n pseudo random numbers from a normal distribution with mean 'mu' and
standard deviation 'sd'}
\usage{
normal(n = 1, mu = 0, sd = 1, seed = as.integer(Sys.time()))
}
\arguments{
\item{n}{the number of pseudo random normal variants to generate.
Defaults to 1.}

\item{mu}{mean of normal distribution. Defaults to 0.}

\item{sd}{standard deviation of normal distribution. Defaults to 1.}

\item{seed}{Seed for pseudo random number generating algorithm. Defaults to a
integer generated from the system time}
}
\value{
vector of PRNs from specified normal distribution of length n
}
\description{
Generate n pseudo random numbers from a normal distribution with mean 'mu' and
standard deviation 'sd'
}
\examples{
normal(n=1e6, mu=0, sd = 3, seed = 122)
}
