% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonSim.R
\name{poissonSim}
\alias{poissonSim}
\title{Generate n pseudo random numbers from a Poisson distribution.}
\usage{
poissonSim(n = 1, lambda = 1, seed = as.integer(Sys.time()))
}
\arguments{
\item{n}{the number of pseudo random Poisson numbers to generate.
Defaults to 1.}

\item{lambda}{Shape parameter indicating the average number off events
in a given time interval. Defaults to 1.}

\item{seed}{Seed for pseudo random number generating algorithm. Defaults to a
integer generated from the system time}
}
\value{
vector of Poisson random variants of length n
}
\description{
Generate n pseudo random numbers from a Poisson distribution.
}
\examples{
poissonSim(n =1e4, lambda=4)
}
