% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang.R
\name{triang}
\alias{triang}
\title{Generate n pseudo random numbers from a triangular distribution starting at
'min', having an expected value of 'peak', and a maximum at
'max'.}
\usage{
triang(n = 1, min = 0, peak = 1, max = 2, seed = as.integer(Sys.time()))
}
\arguments{
\item{n}{the number of pseudo triangular variants to generate.
Defaults to 1.}

\item{min}{Lower bound of distribution. Defaults to 0.}

\item{peak}{Expected value of distribution. Defaults to 1.}

\item{max}{Upper bound of distribution. Defaults to 2.}

\item{seed}{Seed for pseudo random number generating algorithm. Defaults to a
integer generated from the system time}
}
\value{
vector of PRNs from specified triangular distribution of length n
}
\description{
Generate n pseudo random numbers from a triangular distribution starting at
'min', having an expected value of 'peak', and a maximum at
'max'.
}
\examples{
triang(n=1e5, min = 2, peak = 6, max = 10)
}
