% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unif.R
\name{unif}
\alias{unif}
\title{Generate n pseudo random Uniform(0,1) numbers}
\usage{
unif(n = 1, type = "bfs1987", seed = as.integer(Sys.time()))
}
\arguments{
\item{n}{the number of pseudo random Uniform(0,1) numbers to generate.
Defaults to 1.}

\item{type}{Method used. Either "lecuyer" or "bfs1987". Defaults to "lecuyer"}

\item{seed}{Seed for pseudo random number generating algorithm. Defaults to a
integer generated from the system time}
}
\value{
vector of PRN Uniform(0,1) numbers of length n
}
\description{
Generate n pseudo random Uniform(0,1) numbers
}
\examples{
unif(n = 1e5,seed = 12563)
unif(n = 1e5, type = "lecuyer")
}
