% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifab.R
\name{unifab}
\alias{unifab}
\title{Generate n pseudo random Uniform(a,b) numbers between a and b}
\usage{
unifab(n = 1, a = 0, b = 1, type = "lecuyer", seed = as.integer(Sys.time()))
}
\arguments{
\item{n}{the number of pseudo random Uniform(a,b) numbers to generate.
Defaults to 1.}

\item{a}{Lower bound of distribution. Defaults to 0.}

\item{b}{Upper bound of distribution. Defaults to 1.}

\item{type}{Method used. Either "lecuyer" or "bfs1987". Defaults to "lecuyer"}

\item{seed}{Seed for pseudo random number generating algorithm. Defaults to a
integer generated from the system time}
}
\value{
vector of PRN Uniform(a,b) numbers of length n
}
\description{
Generate n pseudo random Uniform(a,b) numbers between a and b
}
\examples{
unifab(n=1e5, a = -3, b = 5)
unifab(n= 1e5, a = 100, b = 200, type = "bfs1987", seed = 1234)
}
