% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{weibull}
\alias{weibull}
\title{Generate n pseudo random numbers from a Weibull distribution.}
\usage{
weibull(n = 1, lambda = 1, k = 1, seed = as.integer(Sys.time()))
}
\arguments{
\item{n}{the number of pseudo random Weibull numbers to generate.
Defaults to 1.}

\item{lambda}{Scale parameter of the distribution. Defaults to 1.}

\item{k}{shape parameter of the distribution. Defaults to 1.}

\item{seed}{Seed for pseudo random number generating algorithm. Defaults to a
integer generated from the system time}
}
\value{
vector of Weibull random variants of length n
}
\description{
Generate n pseudo random numbers from a Weibull distribution.
}
\examples{
weibull(n=1e5, lambda = 1, k = 1)
}
